Refactor the providers and utilities functions
