+++ src/providers/insert-unmock.ts
@@ -1,11 +1,6 @@
 import * as vscode from "vscode";
-import { IInsertUnmockAction } from "./interfaces";
-
-const JS_SUFFIXES = "{js,ts,tsx,jsx}"; // es,es6,ts.erb?
-export const TestJSFilter: vscode.DocumentFilter = {scheme: "file", pattern: `**/*.test.*${JS_SUFFIXES}`}; // containing ".test." in filename
-export const TestJSFolderFilter: vscode.DocumentFilter = {scheme: "file", pattern: `**/test/*.${JS_SUFFIXES}`}; // under "test" folder
-export const TestsJSFolderFilter: vscode.DocumentFilter = {scheme: "file", pattern: `**/tests/*.${JS_SUFFIXES}`}; // under "tests" folder
-export const AllJSFileFilters = [TestJSFilter, TestJSFolderFilter, TestsJSFolderFilter];
+import { IInsertUnmockAction } from "../interfaces";
+import { removeJSCommentsFromSourceText, getRangeFromTextAndMatch } from "../utils";
 
 export class InsertUnmockHoverProvider implements vscode.HoverProvider {
     provideHover(document: vscode.TextDocument,
@@ -95,10 +90,6 @@ function buildTypescriptUnmockActionObject(srcText: string): IInsertUnmockAction
     };
 }
 
-function removeJSCommentsFromSourceText(srcText: string): string {
-    return srcText.replace(/\/\*[\s\S]*?\*\/|([^\\:]|^)\/\/.*$/gm, "");
-}
-
 function unmockJSImportLocation(srcText: string): vscode.Range {
     const matchCall = srcText.match(/^import .*?"unmock".*?$/m);
     if (matchCall === null) {
@@ -117,21 +108,6 @@ function matchJSRequestWithoutUnmock(srcText: string): null | RegExpMatchArray {
             return srcText.match(/[\w_]+\.(?:request|get|delete|head|options|post|put|patch)\(/g);
     }
     return null;
-}
-
-function getRangeFromTextAndMatch(srcText: string, matchCall: RegExpMatchArray): vscode.Range[] {
-    let lastPosition = 0;
-    const splittedSrcText = srcText.split("\n");
-    return matchCall.map((match) => {
-        // -1 for zero-based
-        const pos = srcText.indexOf(match, lastPosition);
-        lastPosition += pos;
-        const lineNumber = srcText.substr(0, pos).split("\n").length - 1;
-        const relevantLine = splittedSrcText[lineNumber];
-        const lineLength = relevantLine.length;
-        const firstCharInLine = relevantLine.length - relevantLine.trimLeft().length;
-        return new vscode.Range(lineNumber, firstCharInLine, lineNumber, lineLength);
-    });
 }
 
 function findLastJSImport(srcText: string): vscode.Position {
